+++ src/rapidfuzz/process.py
@@ -7,10 +7,4 @@ _mod = "rapidfuzz.process"
 extract = _fallback_import(_mod, "extract")
 extractOne = _fallback_import(_mod, "extractOne")
 extract_iter = _fallback_import(_mod, "extract_iter")
-
-try:
-    cdist = _fallback_import("rapidfuzz.process_cdist", "cdist")
-except ImportError:
-
-    def cdist(*args, **kwargs):
-        raise NotImplementedError("implementation requires numpy to be installed")
+cdist = _fallback_import("rapidfuzz.process_cdist", "cdist")