+++ src/rapidfuzz/process_cdist_cpp_impl.pyx
@@ -12,6 +12,8 @@ from libc.stdint cimport uint32_t, uint64_t, int64_t
 
 cimport cython
 from cython.operator cimport dereference
+from cpython cimport Py_buffer
+from cpython.buffer cimport PyBUF_ND, PyBUF_SIMPLE, PyBUF_F_CONTIGUOUS
 
 from cpp_common cimport (
     PyObjectWrapper, RF_StringWrapper, RF_KwargsWrapper,
@@ -20,10 +22,6 @@ from cpp_common cimport (
 )
 
 from array import array
-import numpy as np
-cimport numpy as np
-
-from numpy cimport npy_intp, PyArray_SimpleNew, PyArrayObject
 
 from rapidfuzz_capi cimport (
     RF_Preprocess, RF_Kwargs, RF_String, RF_Scorer, RF_ScorerFunc,
@@ -34,15 +32,48 @@ from rapidfuzz_capi cimport (
 from cpython.pycapsule cimport PyCapsule_IsValid, PyCapsule_GetPointer
 from cpython.object cimport PyObject
 
-np.import_array()
-
 cdef extern from "process_cdist_cpp.hpp":
-    object cdist_single_list_impl[T](  const RF_Kwargs*, RF_Scorer*,
-        const vector[RF_StringWrapper]&, int, int, T) except +
-    object cdist_two_lists_impl[T](    const RF_Kwargs*, RF_Scorer*,
-        const vector[RF_StringWrapper]&, const vector[RF_StringWrapper]&, int, int, T) except +
+    cpdef enum class MatrixType:
+        UNDEFINED = 0
+        FLOAT32 = 1
+        FLOAT64 = 2
+        INT8 = 3
+        INT16 = 4
+        INT32 = 5
+        INT64 = 6
+        UINT8 = 7
+        UINT16 = 8
+        UINT32 = 9
+        UINT64 = 10
 
-    void set_score(PyArrayObject*, int, npy_intp, npy_intp, double)
+    cdef cppclass RfMatrix "Matrix":
+        RfMatrix() except +
+        RfMatrix(MatrixType, size_t, size_t) except +
+        int get_dtype_size() except +
+        const char* get_format() except +
+        void set[T](size_t, size_t, T) except +
+
+        MatrixType m_dtype
+        size_t m_rows
+        size_t m_cols
+        void* m_matrix
+
+    RfMatrix cdist_single_list_impl[T](  const RF_Kwargs*, RF_Scorer*,
+        const vector[RF_StringWrapper]&, MatrixType, int, T) except +
+    RfMatrix cdist_two_lists_impl[T](    const RF_Kwargs*, RF_Scorer*,
+        const vector[RF_StringWrapper]&, const vector[RF_StringWrapper]&, MatrixType, int, T) except +
+
+
+FLOAT32 = MatrixType.FLOAT32
+FLOAT64 = MatrixType.FLOAT64
+INT8 = MatrixType.INT8
+INT16 = MatrixType.INT16
+INT32 = MatrixType.INT32
+INT64 = MatrixType.INT64
+UINT8 = MatrixType.UINT8
+UINT16 = MatrixType.UINT16
+UINT32 = MatrixType.UINT32
+UINT64 = MatrixType.UINT64
 
 cdef inline vector[PyObjectWrapper] preprocess_py(queries, processor) except *:
     cdef vector[PyObjectWrapper] proc_queries
@@ -91,80 +122,88 @@ cdef inline vector[RF_StringWrapper] preprocess(queries, processor) except *:
 
     return move(proc_queries)
 
-cdef inline int dtype_to_type_num(dtype) except -1:
-    if dtype is np.int32:
-        return np.NPY_INT32
-    if dtype is np.int8:
-        return np.NPY_INT8
-    if dtype is np.int16:
-        return np.NPY_INT16
-    if dtype is np.int64:
-        return np.NPY_INT64
-    if dtype is np.uint8:
-        return np.NPY_UINT8
-    if dtype is np.uint16:
-        return np.NPY_UINT16
-    if dtype is np.uint32:
-        return np.NPY_UINT32
-    if dtype is np.uint64:
-        return np.NPY_UINT64
-    if dtype is np.float32:
-        return np.NPY_FLOAT32
-    if dtype is np.float64:
-        return np.NPY_FLOAT64
-
-    raise TypeError("unsupported dtype")
-
-cdef inline int dtype_to_type_num_f64(dtype) except -1:
+cdef inline MatrixType dtype_to_type_num_f64(dtype) except MatrixType.UNDEFINED:
     if dtype is None:
-        return np.NPY_FLOAT32
-    return dtype_to_type_num(dtype)
+        return MatrixType.FLOAT32
+    return <MatrixType>dtype
 
-cdef inline int dtype_to_type_num_i64(dtype) except -1:
+cdef inline MatrixType dtype_to_type_num_i64(dtype) except MatrixType.UNDEFINED:
     if dtype is None:
-        return np.NPY_INT32
-    return dtype_to_type_num(dtype)
+        return MatrixType.INT32
+    return <MatrixType>dtype
+
+from cpython cimport Py_buffer
+from libcpp.vector cimport vector
+
+cdef class Matrix:
+    cdef Py_ssize_t shape[2]
+    cdef RfMatrix matrix
+
+    def __getbuffer__(self, Py_buffer *buffer, int flags):
+        self.shape[0] = self.matrix.m_rows
+        self.shape[1] = self.matrix.m_cols
+
+        buffer.buf = <char *>self.matrix.m_matrix
+        buffer.format = <char *>self.matrix.get_format()
+        buffer.internal = NULL
+        buffer.itemsize = self.matrix.get_dtype_size()
+        buffer.len = self.matrix.m_rows * self.matrix.m_cols * self.matrix.get_dtype_size()
+        buffer.ndim = 2
+        buffer.obj = self
+        buffer.readonly = 0
+        buffer.shape = self.shape
+        buffer.strides = NULL
+        buffer.suboffsets = NULL
+
+    def __releasebuffer__(self, Py_buffer *buffer):
+        pass
 
 cdef cdist_two_lists(queries, choices, RF_Scorer* scorer, const RF_ScorerFlags* scorer_flags, processor, score_cutoff, dtype, int c_workers, const RF_Kwargs* kwargs):
     proc_queries = preprocess(queries, processor)
     proc_choices = preprocess(choices, processor)
     flags = dereference(scorer_flags).flags
+    cdef Matrix matrix = Matrix()
 
     if flags & RF_SCORER_FLAG_RESULT_F64:
-        return cdist_two_lists_impl(
+        matrix.matrix = cdist_two_lists_impl(
             kwargs, scorer, proc_queries, proc_choices,
             dtype_to_type_num_f64(dtype),
             c_workers,
             get_score_cutoff_f64(score_cutoff, scorer_flags))
 
     elif flags & RF_SCORER_FLAG_RESULT_I64:
-        return cdist_two_lists_impl(
+        matrix.matrix = cdist_two_lists_impl(
             kwargs, scorer, proc_queries, proc_choices,
             dtype_to_type_num_i64(dtype),
             c_workers,
             get_score_cutoff_i64(score_cutoff, scorer_flags))
+    else:
+        raise ValueError("scorer does not properly use the C-API")
 
-    raise ValueError("scorer does not properly use the C-API")
+    return matrix
 
-cdef cdist_single_list(queries, RF_Scorer* scorer, const RF_ScorerFlags* scorer_flags, processor, score_cutoff, dtype, int c_workers, const RF_Kwargs* kwargs):
+cdef Matrix cdist_single_list(queries, RF_Scorer* scorer, const RF_ScorerFlags* scorer_flags, processor, score_cutoff, dtype, int c_workers, const RF_Kwargs* kwargs):
     proc_queries = preprocess(queries, processor)
     flags = dereference(scorer_flags).flags
+    cdef Matrix matrix = Matrix()
 
     if flags & RF_SCORER_FLAG_RESULT_F64:
-        return cdist_single_list_impl(
+        matrix.matrix = cdist_single_list_impl(
             kwargs, scorer, proc_queries,
             dtype_to_type_num_f64(dtype),
             c_workers,
             get_score_cutoff_f64(score_cutoff, scorer_flags))
 
     elif flags & RF_SCORER_FLAG_RESULT_I64:
-        return cdist_single_list_impl(
+        matrix.matrix = cdist_single_list_impl(
             kwargs, scorer, proc_queries,
             dtype_to_type_num_i64(dtype),
             c_workers,
             get_score_cutoff_i64(score_cutoff, scorer_flags))
+    else:
+        raise ValueError("scorer does not properly use the C-API")
 
-    raise ValueError("scorer does not properly use the C-API")
+    return matrix
 
 
 @cython.boundscheck(False)
@@ -172,11 +211,10 @@ cdef cdist_single_list(queries, RF_Scorer* scorer, const RF_ScorerFlags* scorer_
 cdef cdist_py(queries, choices, scorer, processor, score_cutoff, dtype, workers, dict kwargs):
     proc_queries = preprocess_py(queries, processor)
     proc_choices = preprocess_py(choices, processor)
-
-    cdef npy_intp[2] dims = [<npy_intp>proc_queries.size(), <npy_intp>proc_choices.size()]
-
+    cdef double score
+    cdef Matrix matrix = Matrix()
     c_dtype = dtype_to_type_num_f64(dtype)
-    matrix = PyArray_SimpleNew(2, dims, c_dtype)
+    matrix.matrix = RfMatrix(c_dtype, proc_queries.size(), proc_choices.size())
 
     kwargs["processor"] = None
     kwargs["score_cutoff"] = score_cutoff
@@ -184,64 +222,12 @@ cdef cdist_py(queries, choices, scorer, processor, score_cutoff, dtype, workers,
     for i in range(proc_queries.size()):
         for j in range(proc_choices.size()):
             score = scorer(<object>proc_queries[i].obj, <object>proc_choices[j].obj,**kwargs)
-            set_score(<PyArrayObject*>matrix, c_dtype, i, j, score)
+            matrix.matrix.set(i, j, score)
 
     return matrix
 
-# todo link to c api docs
+
 def cdist(queries, choices, *, scorer=ratio, processor=None, score_cutoff=None, dtype=None, workers=1, **kwargs):
-    """
-    Compute distance/similarity between each pair of the two collections of inputs.
-
-    Parameters
-    ----------
-    queries : Collection[Sequence[Hashable]]
-        list of all strings the queries
-    choices : Collection[Sequence[Hashable]]
-        list of all strings the query should be compared
-    scorer : Callable, optional
-        Optional callable that is used to calculate the matching score between
-        the query and each choice. This can be:
-
-        - a scorer using the RapidFuzz C-API like the builtin scorers in RapidFuzz,
-          which can return a distance or similarity between two strings. Further details can be found here.
-        - a Python function which returns a similarity between two strings in the range 0-100. This is not
-          recommended, since it is far slower than a scorer using the RapidFuzz C-API.
-
-        fuzz.ratio is used by default.
-    processor : Callable, optional
-        Optional callable that is used to preprocess the strings before
-        comparing them. Default is None, which deactivates this behaviour.
-    score_cutoff : Any, optional
-        Optional argument for a score threshold to be passed to the scorer.
-        Default is None, which deactivates this behaviour.
-    dtype : data-type, optional
-        The desired data-type for the result array.Depending on the scorer type the following
-        dtypes are supported:
-
-        - similarity:
-          - np.float32, np.float64
-          - np.uint8 -> stores fixed point representation of the result scaled to a range 0-100
-        - distance:
-          - np.int8, np.int16, np.int32, np.int64
-
-        If not given, then the type will be np.float32 for similarities and np.int32 for distances.
-    workers : int, optional
-        The calculation is subdivided into workers sections and evaluated in parallel.
-        Supply -1 to use all available CPU cores.
-        This argument is only available for scorers using the RapidFuzz C-API so far, since it
-        releases the Python GIL.
-    **kwargs : Any, optional
-        any other named parameters are passed to the scorer. This can be used to pass
-        e.g. weights to string_metric.levenshtein
-
-    Returns
-    -------
-    ndarray
-        Returns a matrix of dtype with the distance/similarity between each pair
-        of the two collections of inputs.
-    """
-
     cdef RF_Scorer* scorer_context = NULL
     cdef RF_ScorerFlags scorer_flags